requirements.txt
browser-use
langchain-google-genai~=2.1.2
playwright
pydantic~=2.11.1
python-dotenv~=1.1.0
pyyaml
Jinja2~=3.1.6
openai~=1.70.0
pytest~=8.3.5